hw_5
a program that takes the path to the file as the first argument to the command line, and the file name as the second.
CREATE file and directory if doesn't exist. Print list of files with metadata such as file sizes, modification dates for any directory which is put as first argument.